refactor(types): split audit-log.ts into entities/audit-log-entry.ts
