+++ packages/types/src/__tests__/audit-log.test.ts
@@ -1,7 +1,7 @@
 import { assertType, describe, expectTypeOf, it } from "vitest";
 
-import type { AuditEventType, AuditLogEntry } from "../audit-log.js";
 import type { Plaintext } from "../encryption.js";
+import type { AuditEventType, AuditLogEntry } from "../entities/audit-log-entry.js";
 import type { AccountId, ApiKeyId, AuditLogEntryId, SystemId } from "../ids.js";
 import type { UnixMillis } from "../timestamps.js";
