+++ packages/types/src/__tests__/sot-manifest.test.ts
@@ -1,13 +1,13 @@
 import { describe, expectTypeOf, it } from "vitest";
 
 import type { SotEntityManifest } from "../__sot-manifest__.js";
-import type { AuditLogEntry } from "../audit-log.js";
 import type {
   AuditLogEntryServerMetadata,
   AuditLogEntryWire,
   MemberServerMetadata,
   MemberWire,
 } from "../encryption.js";
+import type { AuditLogEntry } from "../entities/audit-log-entry.js";
 import type { Member } from "../entities/member.js";
 import type { Equal, Extends } from "../type-assertions.js";
