+++ packages/types/src/encryption.ts
@@ -1,6 +1,6 @@
-import type { AuditEventType, AuditActor, AuditLogEntry } from "./audit-log.js";
 import type { KdfMasterKey } from "./crypto-keys.js";
 import type { AcknowledgementRequest } from "./entities/acknowledgement.js";
+import type { AuditEventType, AuditActor, AuditLogEntry } from "./entities/audit-log-entry.js";
 import type { BoardMessage } from "./entities/board-message.js";
 import type { Channel } from "./entities/channel.js";
 import type { CustomFront } from "./entities/custom-front.js";
