+++ packages/types/src/entities/audit-log-entry.ts
@@ -1,6 +1,6 @@
-import type { Plaintext } from "./encryption.js";
-import type { AccountId, ApiKeyId, AuditLogEntryId, SystemId } from "./ids.js";
-import type { UnixMillis } from "./timestamps.js";
+import type { Plaintext } from "../encryption.js";
+import type { AccountId, ApiKeyId, AuditLogEntryId, SystemId } from "../ids.js";
+import type { UnixMillis } from "../timestamps.js";
 
 /** Step names used in setup wizard audit events. */
 export type SetupStepName = "nomenclature" | "profile";
